Added star counter in the sidebar, and a lifebar at sidebar
