+++ game.kv
@@ -21,15 +21,31 @@ ScreenManager:
     on_leave: app.screen_on_leave(1)
     BoxLayout:
         orientation: "horizontal"
-        ToggleButton:
-            id: kiss_button_lvl1
-            size_hint: app.side_bar_width, 0.15
-            text: "Kiss"
-            on_state: app.on_toggle_button_state(self, 1)
-            #size_hint: 0.15, 0.15
-            #height: "90dp"
-            #width: "90dp"
-            #pos_hint: {'x': 0.45, 'y': 0.1}
+        BoxLayout:
+            orientation: "vertical"
+            spacing: 2
+            size_hint: app.side_bar_width, 1
+            RemainingLifePercent
+                id: remaining_life_percent_lvl1
+            BoxLayout:
+                size_hint: 1, 0.03
+                Label:
+                    canvas:
+                        Rectangle:
+                            pos: self.pos
+                            size: self.size
+                            source: "graphics/entities/star_small.png"
+                NumCollectedStars:
+                    id: num_stars_collected_lvl1
+            ToggleButton:
+                id: kiss_button_lvl1
+                size_hint: 1, 0.2
+                text: "Kiss"
+                on_state: app.on_toggle_button_state(self, 1)
+                #size_hint: 0.15, 0.15
+                #height: "90dp"
+                #width: "90dp"
+                #pos_hint: {'x': 0.45, 'y': 0.1}
         FloatLayout:
             #size_hint: 0, 0.9
             id: layout_lvl1
@@ -39,8 +55,6 @@ ScreenManager:
                     size: self.size
                     pos: self.pos
                     source: "graphics/levels-bg/PvZ_day_bg_lvl1_cropped.jpg"
-            NumCollectedStars:
-                id: num_stars_collected_lvl1
             Character:
                 id: character_image_lvl1
                 pos: root.center_x-self.size[0]/2, root.center_y-self.size[1]/2
@@ -55,15 +69,31 @@ ScreenManager:
     on_leave: app.screen_on_leave(2)
     BoxLayout:
         orientation: "horizontal"
-        ToggleButton:
-            id: kiss_button_lvl2
-            size_hint: app.side_bar_width, 0.15
-            text: "Kiss"
-            on_state: app.on_toggle_button_state(self, 2)
-            #size_hint: 0.15, 0.15
-            #height: "90dp"
-            #width: "90dp"
-            #pos_hint: {'x': 0.45, 'y': 0.1}
+        BoxLayout:
+            orientation: "vertical"
+            spacing: 2
+            size_hint: app.side_bar_width, 1
+            RemainingLifePercent
+                id: remaining_life_percent_lvl1
+            BoxLayout:
+                size_hint: 1, 0.03
+                Label:
+                    canvas:
+                        Rectangle:
+                            pos: self.pos
+                            size: self.size
+                            source: "graphics/entities/star_small.png"
+                NumCollectedStars:
+                    id: num_stars_collected_lvl1
+            ToggleButton:
+                id: kiss_button_lvl1
+                size_hint: 1, 0.2
+                text: "Kiss"
+                on_state: app.on_toggle_button_state(self, 1)
+                #size_hint: 0.15, 0.15
+                #height: "90dp"
+                #width: "90dp"
+                #pos_hint: {'x': 0.45, 'y': 0.1}
         FloatLayout:
             id: layout_lvl2
             on_touch_down: app.touch_down_handler(2, args)
@@ -79,11 +109,14 @@ ScreenManager:
                 pos: root.center_x-self.size[0]/2, root.center_y-self.size[1]/2
 
 <NumCollectedStars@Label>:
-    id: num_stars_collected
-    size_hint: (0.18, 0.08)
-    pos_hint: {'x': 0.0, 'y': 0.90}
-    text: "Stars: 0/0"
+    text: "0/0"
     font_size: 20
+    #background_normal: ''
+    #id: num_stars_collected
+    #size_hint: (1, 0.15)
+    #pos_hint: {'x': 0.0, 'y': 0.20}
+
+
 
 #<Monster@Image>:
 #    id: monster_image
@@ -106,3 +139,15 @@ ScreenManager:
     #on_pos_hint: app.char_pos_hint(self)
     #start_im_num: 0
     #end_im_num: 7
+
+<RemainingLifePercent@Label>:
+    #remaining_life_size_hint_y: 0.3
+    size_hint: (1, 0.3)
+    #pos_hint: {'x': 0.0, 'y': 0.4}
+    #font_size: 20
+    canvas:
+        Color:
+            rgb: (0, 0, 1)
+        Rectangle:
+            pos: self.pos
+            size: self.size
